[refactor/#170] EditMainScreen 컴포넌트 분리 및 Preview 추가
