+++ app/src/main/java/com/bongtu/baekseo/presentation/edit/EditMainScreen.kt
@@ -1,21 +1,15 @@
 package com.bongtu.baekseo.presentation.edit
 
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.ime
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -25,27 +19,23 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
 import com.bongtu.baekseo.R.drawable.ic_arrow_back
-import com.bongtu.baekseo.R.drawable.ic_arrow_down
-import com.bongtu.baekseo.R.drawable.ic_arrow_up
 import com.bongtu.baekseo.R.drawable.ic_calendar
 import com.bongtu.baekseo.R.drawable.ic_check_gray
 import com.bongtu.baekseo.R.drawable.ic_coin
@@ -71,7 +61,6 @@ import com.bongtu.baekseo.R.string.edit_nickname_text_field_placeholder
 import com.bongtu.baekseo.R.string.edit_nickname_title
 import com.bongtu.baekseo.R.string.edit_relation_dropdown_placeholder
 import com.bongtu.baekseo.R.string.edit_relation_title
-import com.bongtu.baekseo.R.string.edit_required_text
 import com.bongtu.baekseo.R.string.edit_save_button
 import com.bongtu.baekseo.R.string.kr_won
 import com.bongtu.baekseo.core.common.type.AttendType
@@ -83,7 +72,6 @@ import com.bongtu.baekseo.core.common.type.TopBarType
 import com.bongtu.baekseo.core.compositionlocal.safeDrawingWithBottomNavBar
 import com.bongtu.baekseo.core.designsystem.component.button.BongBaekButton
 import com.bongtu.baekseo.core.designsystem.component.dialog.BongBaekDatePickerDialog
-import com.bongtu.baekseo.core.designsystem.component.dropdownmenu.BongBaekDropdownMenu
 import com.bongtu.baekseo.core.designsystem.component.textfield.LabelTextField
 import com.bongtu.baekseo.core.designsystem.component.topbar.BongBaekTopBar
 import com.bongtu.baekseo.core.designsystem.theme.BongBaekTheme
@@ -100,10 +88,11 @@ import com.bongtu.baekseo.presentation.edit.EditContract.EditSideEffect.EditMain
 import com.bongtu.baekseo.presentation.edit.EditContract.EditSideEffect.EditMainSideEffect.NavigateToRecord
 import com.bongtu.baekseo.presentation.edit.EditContract.EditSideEffect.EditMainSideEffect.NavigateToSchedule
 import com.bongtu.baekseo.presentation.edit.EditContract.EditUiState
+import com.bongtu.baekseo.presentation.edit.component.EditDropdownMenu
 import com.bongtu.baekseo.presentation.edit.component.EditLocationContent
 import com.bongtu.baekseo.presentation.edit.component.EditMemoContent
+import com.bongtu.baekseo.presentation.edit.component.EditFieldItem
 import com.bongtu.baekseo.presentation.edit.type.EditEntryType
-import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.flow.filterIsInstance
 
@@ -257,11 +246,11 @@ private fun EditMainScreen(
                     isEditable = isResultEditable,
                 )
 
-                FormFieldItem(
+                EditFieldItem(
                     iconRes = ic_relation,
                     labelRes = edit_relation_title,
                     content = {
-                        FormFieldDropDown(
+                        EditDropdownMenu(
                             placeholder = stringResource(id = edit_relation_dropdown_placeholder),
                             menuItems = relations,
                             selectedItem = uiState.relationship,
@@ -272,11 +261,11 @@ private fun EditMainScreen(
                     modifier = Modifier.padding(bottom = 32.dp),
                 )
 
-                FormFieldItem(
+                EditFieldItem(
                     iconRes = ic_event,
                     labelRes = edit_event_title,
                     content = {
-                        FormFieldDropDown(
+                        EditDropdownMenu(
                             placeholder = stringResource(id = edit_event_dropdown_placeholder),
                             menuItems = events,
                             selectedItem = uiState.eventCategory,
@@ -319,11 +308,11 @@ private fun EditMainScreen(
                     )
                 }
 
-                FormFieldItem(
+                EditFieldItem(
                     iconRes = ic_check_gray,
                     labelRes = edit_is_attend_title,
                     content = {
-                        FormFieldDropDown(
+                        EditDropdownMenu(
                             placeholder = stringResource(id = edit_is_attend_dropdown_placeholder),
                             menuItems = attendOptions,
                             selectedItem = uiState.attendLabel,
@@ -351,7 +340,7 @@ private fun EditMainScreen(
                     visualTransformation = DateTextFieldFormat(),
                 )
 
-                FormFieldItem(
+                EditFieldItem(
                     iconRes = ic_location,
                     labelRes = edit_location_title,
                     content = {
@@ -418,126 +407,25 @@ private fun EditMainScreen(
     }
 }
 
+@Preview
 @Composable
-private fun FormFieldItem(
-    @DrawableRes iconRes: Int,
-    @StringRes labelRes: Int,
-    content: @Composable () -> Unit,
-    modifier: Modifier = Modifier,
-    isRequired: Boolean = true,
-    trailing: (@Composable () -> Unit)? = null,
-) {
-    Column(
-        modifier = modifier,
-    ) {
-        Row(
-            modifier = Modifier
-                .padding(bottom = 12.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(
-                imageVector = ImageVector.vectorResource(id = iconRes),
-                contentDescription = null,
-                modifier = Modifier
-                    .height(16.dp),
-                tint = Color.Unspecified,
-            )
-
-            Spacer(modifier = Modifier.width(6.dp))
-
-            Text(
-                text = stringResource(id = labelRes),
-                style = BongBaekTheme.typography.body1Medium14,
-                color = BongBaekTheme.colors.white,
-            )
-
-            if (isRequired) {
-                Spacer(modifier = Modifier.width(2.dp))
-                Text(
-                    text = stringResource(id = edit_required_text),
-                    style = BongBaekTheme.typography.body1Medium14,
-                    color = BongBaekTheme.colors.primaryNormal,
-                )
-            }
-
-            Spacer(modifier = Modifier.weight(1f))
-
-            trailing?.invoke()
-        }
-        content.invoke()
-    }
-}
-
-@Composable
-private fun FormFieldDropDown(
-    placeholder: String,
-    menuItems: ImmutableList<String>,
-    selectedItem: String,
-    onItemSelected: (String) -> Unit,
-    isEditable: Boolean = true,
-) {
-    var expanded by remember { mutableStateOf(false) }
-    val isSelected = selectedItem.isNotBlank()
-    val bongBaekColors = BongBaekTheme.colors
-    val text = if (isSelected) selectedItem else placeholder
-    val textColor = when {
-        isSelected && expanded -> bongBaekColors.primaryNormal
-        isSelected || expanded -> bongBaekColors.white
-        else -> bongBaekColors.gray300
-    }
-    val borderColor =
-        if (expanded) BongBaekTheme.colors.primaryNormal else BongBaekTheme.colors.transparent
-    var rowWidthPx by remember { mutableIntStateOf(0) }
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(
-                color = BongBaekTheme.colors.gray750,
-                shape = RoundedCornerShape(10.dp),
-            )
-            .onGloballyPositioned { coordinates ->
-                rowWidthPx = coordinates.size.width
-            }
-            .border(
-                width = 1.dp,
-                color = borderColor,
-                shape = RoundedCornerShape(10.dp),
-            )
-            .padding(
-                vertical = 12.dp,
-                horizontal = 16.dp,
-            )
-            .noRippleClickable {
-                if (isEditable && !expanded) expanded = true
-            },
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Text(
-            text = text,
-            style = BongBaekTheme.typography.body1Medium16,
-            color = textColor,
-        )
-
-        Spacer(modifier = Modifier.weight(1f))
-
-        Icon(
-            imageVector = ImageVector.vectorResource(id = if (expanded) ic_arrow_up else ic_arrow_down),
-            contentDescription = null,
-            tint = textColor,
+private fun EditMainScreenPreview() {
+    BongBaekTheme {
+        EditMainScreen(
+            editEntryType = EditEntryType.FROM_DETAIL,
+            uiState = EditUiState(),
+            navigateUp = {},
+            navigateToLocation = {},
+            onNameChange = {},
+            onNicknameChange = {},
+            onRelationSelect = {},
+            onEventSelect = {},
+            onCostChange = {},
+            onAttendSelect = {},
+            onDateChange = {},
+            onNoteChange = {},
+            checkIsFormFilled = { true },
+            onSubmitEventButtonClick = {},
         )
     }
-
-    BongBaekDropdownMenu(
-        expanded = expanded,
-        items = menuItems,
-        selectedItem = selectedItem,
-        onDismissRequest = { expanded = false },
-        onItemSelect = { onItemSelected(it) },
-        label = { it },
-        modifier = Modifier
-            .width(with(LocalDensity.current) { rowWidthPx.toDp() }),
-        maxItemSize = menuItems.size,
-        isFocusable = true,
-    )
 }
